backend/requirements-dev.txt
-r dependencies/requirements.txt

pytest==7.1.2
pytest-mock==3.8.2
moto==4.0.1
pytest-flake8==1.1.1
pytest-watch==4.2.0
pytest-cov==3.0.0
pytest-timeout==2.1.0
pytest-order==1.0.1
freezegun==1.2.2
python-gnupg==0.5.0
